docs: update chip doc
