+++ src/Uno.Toolkit.UI/Controls/Chips/Chip.cs
@@ -20,115 +20,6 @@ namespace Uno.Toolkit.UI
 	{
 		private const string RemoveButtonName = "PART_RemoveButton";
 
-		public event ChipRemovingEventHandler Removing;
-		public event RoutedEventHandler Removed;
-
-		/// <summary>
-		/// Fires when a ToggleButton is checked or unchecked, except when set with <see cref="SetIsCheckedSilently"/>.
-		/// </summary>
-		internal event RoutedEventHandler IsCheckedChanged;
-
-#region DependencyProperty: IsCheckable = true
-
-		public static DependencyProperty IsCheckableProperty { get; } = DependencyProperty.Register(
-			nameof(IsCheckable),
-			typeof(bool),
-			typeof(Chip),
-			new PropertyMetadata(true, (s, e) => (s as Chip)?.OnIsCheckableChanged(e)));
-
-		/// <summary>
-		/// Gets or sets whether this should behave like a ToggleButton or a Button
-		/// </summary>
-		public bool IsCheckable
-		{
-			get => (bool)GetValue(IsCheckableProperty);
-			set => SetValue(IsCheckableProperty, value);
-		}
-
-#endregion
-
-#region DependencyProperty: Icon
-
-		public object Icon
-		{
-			get { return (IconElement)GetValue(IconProperty); }
-			set { SetValue(IconProperty, value); }
-		}
-
-		public static readonly DependencyProperty IconProperty =
-			DependencyProperty.Register("Icon", typeof(object), typeof(Chip), new PropertyMetadata(null));
-
-#endregion
-
-#region DependencyProperty: IconTemplate
-
-		public DataTemplate IconTemplate
-		{
-			get { return (DataTemplate)GetValue(IconTemplateProperty); }
-			set { SetValue(IconTemplateProperty, value); }
-		}
-
-		public static readonly DependencyProperty IconTemplateProperty =
-			DependencyProperty.Register("IconTemplate", typeof(DataTemplate), typeof(Chip), new PropertyMetadata(null));
-
-#endregion
-
-#region DependencyProperty: CanRemove = false
-
-		public bool CanRemove
-		{
-			get { return (bool)GetValue(CanRemoveProperty); }
-			set { SetValue(CanRemoveProperty, value); }
-		}
-
-		public static readonly DependencyProperty CanRemoveProperty =
-			DependencyProperty.Register("CanRemove", typeof(bool), typeof(Chip), new PropertyMetadata(false));
-
-#endregion
-
-#region DependencyProperty: RemovedCommand
-
-		public static DependencyProperty RemovedCommandProperty { get; } = DependencyProperty.Register(
-			nameof(RemovedCommand),
-			typeof(ICommand),
-			typeof(Chip),
-			new PropertyMetadata(default));
-
-		public ICommand RemovedCommand
-		{
-			get => (ICommand)GetValue(RemovedCommandProperty);
-			set => SetValue(RemovedCommandProperty, value);
-		}
-
-#endregion
-
-#region DependencyProperty: RemovedCommandParameter
-
-		public static DependencyProperty RemovedCommandParameterProperty { get; } = DependencyProperty.Register(
-			nameof(RemovedCommandParameter),
-			typeof(object),
-			typeof(Chip),
-			new PropertyMetadata(default));
-
-		public object RemovedCommandParameter
-		{
-			get => (object)GetValue(RemovedCommandParameterProperty);
-			set => SetValue(RemovedCommandParameterProperty, value);
-		}
-
-		#endregion
-
-#region DependencyProperty: Elevation
-		public double Elevation
-		{
-			get { return (double)GetValue(ElevationProperty); }
-			set { SetValue(ElevationProperty, value); }
-		}
-
-		public static readonly DependencyProperty ElevationProperty =
-			DependencyProperty.Register("Elevation", typeof(double), typeof(Chip), new PropertyMetadata(0d));
-#endregion
-
 		private bool _isMuted = false;
 
 		public Chip()
