+++ src/Uno.Toolkit.UI/Controls/Chips/ChipSelectionMode.cs
@@ -6,6 +6,19 @@ namespace Uno.Toolkit.UI
 {
 	public enum ChipSelectionMode
 	{
-		None, Single, Multiple,
+		/// <summary>
+		/// Selection is disabled.
+		/// </summary>
+		None,
+
+		/// <summary>
+		/// Only one item can be selected at a time.
+		/// </summary>
+		Single,
+
+		/// <summary>
+		/// Multiple items can be selected.
+		/// </summary>
+		Multiple,
 	}
 }
