+++ src/Uno.Toolkit.UI/Controls/Chips/Chips.EventHandlers.cs
@@ -6,8 +6,11 @@ namespace Uno.Toolkit.UI
 {
 	#region Chip event handlers
 	public delegate void ChipRemovingEventHandler(object sender, ChipRemovingEventArgs e);
-	public sealed class ChipRemovingEventArgs : EventArgs
+	public class ChipRemovingEventArgs : EventArgs
 	{
+		/// <summary>
+		/// Gets or sets whether the <see cref="Chip.Removed"/> event should be canceled.
+		/// </summary>
 		public bool Cancel { get; set; }
 	}
 
@@ -21,14 +24,20 @@ namespace Uno.Toolkit.UI
 	{
 		internal ChipItemEventArgs(object item) => Item = item;
 
+		/// <summary>
+		/// Gets the item associated with the event
+		/// </summary>
 		public object Item { get; }
 	}
 
-	public delegate void ChipItemRemovingEventHandler(object sender, ChipItemEventArgs e);
+	public delegate void ChipItemRemovingEventHandler(object sender, ChipItemRemovingEventArgs e);
 	public class ChipItemRemovingEventArgs : ChipItemEventArgs
 	{
 		public ChipItemRemovingEventArgs(object item) : base(item) { }
 
+		/// <summary>
+		/// Gets or sets whether the <see cref="ChipGroup.ItemRemoved"/> event should be canceled.
+		/// </summary>
 		public bool Cancel { get; set; }
 	}
